ENH: special: make `ellipk` a ufunc
No algorithmic changes are made; the main point of the change is that
it causes `ellipk` to show up in `cython_special`.
