+++ scipy/special/_ellipk.pxd
@@ -1,0 +1,5 @@
+from ._cephes cimport ellpk
+
+
+cdef inline double ellipk(double m) nogil:
+    return ellpk(1.0 - m)
